+++ packages/engine-canvas/index.js
@@ -7,10 +7,11 @@ export const run = (functionName, func, values, config) => {
 
   root.innerHTML = "";
 
+  let isElAdded = false;
+
   const mechanic = new Mechanic(func.settings, values, config);
 
-  let isElAdded = false;
-  const onFrame = el => {
+  mechanic.registerFrameCallback(el => {
     if (!isElAdded) {
       isElAdded = true;
       root.appendChild(el);
@@ -18,11 +19,9 @@ export const run = (functionName, func, values, config) => {
     if (!isPreview) {
       mechanic.frame(el);
     }
-  };
-  const onDone = async (el, name) => {
-    // TODO: This could go to core, because we always want to stop the drawloop
-    // when we're done drawing.
-    mechanic.drawLoop.stop();
+  });
+
+  mechanic.registerDoneCallback(async (el, name) => {
     if (!isElAdded) {
       isElAdded = true;
       root.appendChild(el);
@@ -31,22 +30,12 @@ export const run = (functionName, func, values, config) => {
       await mechanic.done(el);
       mechanic.download(name || functionName);
     }
-  };
-  const onSetState = async obj => {
-    mechanic.setState(obj);
-  };
-
-  const { drawLoop } = mechanic.callbacksForEngine();
+  });
 
   func.handler({
     inputs: mechanic.values,
-    mechanic: {
-      frame: onFrame,
-      done: onDone,
-      drawLoop,
-      state: mechanic.functionState,
-      setState: onSetState
-    }
+    ...mechanic.callbacksForEngine()
   });
+
   return mechanic;
 };